refactor(weather): migrate OpenWeatherMap provider to server-side
Add server-side OpenWeatherMap provider using HTTPFetcher, following the
same pattern as OpenMeteo migration.

Changes:
- Add openweathermap_server.js with OnceCall API 3.0 support
- Update weather.js to recognize openweathermap as server-side provider
- Remove client-side openweathermap.js (now obsolete)

Features:
- Supports current, forecast/daily, and hourly weather types
- Timezone offset handling for accurate timestamps
- UV index, precipitation, and feels-like temperature
- HTTPFetcher automatic polling and retry strategies
